Create a new command "addwp"
The new command "addwp" is to add the pid and its parnet pid into the
hidden list.

This command can be used in the following form:

  echo "addwp <pid>" | sudo tee /dev/hideproc

If the operation succeed, you can see both the pid and its parent pid
will appear in the hidden list. You can verify it by the following:

  sudo cat /dev/hideproc